refactor(client): HASHI-70 KakaoStartButton shared로 이동
