+++ apps/client/src/shared/components/kakaoStartButton/KakaoStartButton.tsx
@@ -1,6 +1,5 @@
+import { KakaoIcon } from '@hashi/hds-icons'
 import { Button } from '@hashi/hds-ui'
-
-import { KakaoIcon } from '@/features/auth/assets/KakaoIcon'
 
 type KakaoStartButtonProps = {
   children: string
@@ -13,12 +12,11 @@ export const KakaoStartButton = ({
 }: KakaoStartButtonProps) => {
   return (
     <Button
-      className="typo-sub-header-1 bg-point-200 relative h-[58px] rounded-full text-black"
+      className="typo-sub-header-1 bg-point-200 relative h-14.5 w-85 rounded-full text-black"
       leftIcon={
-        <KakaoIcon className="absolute top-1/2 left-[19px] size-6 -translate-y-1/2" />
+        <KakaoIcon className="absolute top-1/2 left-4.75 size-6 -translate-y-1/2" />
       }
       onClick={onPress}
-      width="full"
     >
       {children}
     </Button>
