+++ apps/client/src/shared/components/kakaoStartButton/index.ts
@@ -1,0 +1,1 @@
+export { KakaoStartButton } from './KakaoStartButton'
